fix URI parsing regex
Remove `'` and `)` characters from `urlchar` macro to
properly identify the end of an URI if it is followed by
another URI
(Example: `a{background:url(a.com)}b{background:url(b.com)}`)

`stringchar` is inherited from `urlchar`, so the removed `)` is
added back to `stringchar` macro since it's a valid string character.
`'` skipped, because `string` macro handles it.